listener re-code
