+++ src/main/java/org/cdpg/dx/databroker/client/RabbitClient.java
@@ -13,9 +13,7 @@ import io.vertx.core.buffer.Buffer;
 import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
 import io.vertx.ext.web.client.HttpResponse;
-import io.vertx.rabbitmq.QueueOptions;
 import io.vertx.rabbitmq.RabbitMQClient;
-import io.vertx.rabbitmq.RabbitMQConsumer;
 import io.vertx.serviceproxy.ServiceException;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -626,24 +624,7 @@ public class RabbitClient {
         .onFailure(
             failure -> {
               LOGGER.error("Fail : " + failure.getMessage());
-              promise.fail(failure);
-            });
-    return promise.future();
-  }
-
-  public Future<RabbitMQConsumer> basicConsumeInternal(String queueName, QueueOptions options) {
-    Promise<RabbitMQConsumer> promise = Promise.promise();
-    iudxInternalRabbitMqClient
-        .basicConsumer(queueName, options)
-        .onSuccess(
-            resultHandler -> {
-              LOGGER.info("Success : Consumed");
-              promise.complete(resultHandler);
-            })
-        .onFailure(
-            failure -> {
-              LOGGER.error("Fail : " + failure.getMessage());
-              promise.fail(failure);
+              promise.fail(new ServiceException(ERROR_INTERNAL_SERVER, INTERNAL_SERVER_ERROR));
             });
     return promise.future();
   }
