+++ src/main/java/org/cdpg/dx/databroker/listeners/util/BroadcastEventType.java
@@ -1,4 +1,4 @@
-package org.cdpg.dx.rs.listners.util;
+package org.cdpg.dx.databroker.listeners.util;
 
 import java.util.HashMap;
 import java.util.Map;
