--- src/main/java/org/cdpg/dx/rs/listners/UniqueAttribQlistener.java
@@ -1,83 +1,0 @@
-package org.cdpg.dx.rs.listners;
-
-import static org.cdpg.dx.common.ErrorCode.ERROR_INTERNAL_SERVER;
-import static org.cdpg.dx.common.ErrorMessage.INTERNAL_SERVER_ERROR;
-import static org.cdpg.dx.rs.listners.util.Constans.UNIQUE_ATTR_Q;
-
-import io.vertx.core.Future;
-import io.vertx.core.Promise;
-import io.vertx.core.buffer.Buffer;
-import io.vertx.core.json.JsonObject;
-import io.vertx.rabbitmq.QueueOptions;
-import io.vertx.serviceproxy.ServiceException;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-import org.cdpg.dx.databroker.service.DataBrokerService;
-import org.cdpg.dx.rs.listners.util.BroadcastEventType;
-import org.cdpg.dx.uniqueattribute.service.UniqueAttributeService;
-
-public class UniqueAttribQlistener {
-  private static final Logger LOGGER = LogManager.getLogger(UniqueAttribQlistener.class);
-  private final QueueOptions options =
-      new QueueOptions().setMaxInternalQueueSize(1000).setKeepMostRecent(true);
-  private final DataBrokerService dataBrokerService;
-  private final UniqueAttributeService uniqueAttributeService;
-
-  // TODO: Start from main verticle
-  public UniqueAttribQlistener(
-      DataBrokerService dataBrokerService, UniqueAttributeService uniqueAttributeService) {
-    this.dataBrokerService = dataBrokerService;
-    this.uniqueAttributeService = uniqueAttributeService;
-  }
-
-  public Future<Void> start() {
-    Promise<Void> promise = Promise.promise();
-    LOGGER.info("Unique Attribute Queue listener started");
-    dataBrokerService
-        .basicConsumeInternal(UNIQUE_ATTR_Q, options)
-        .onSuccess(
-            uniqueHandler -> {
-              uniqueHandler.handler(
-                  message -> {
-                    Buffer body = message.body();
-                    if (body != null) {
-                      JsonObject uniqueAttribJson = new JsonObject(body);
-                      LOGGER.debug("received message from unique-attrib Q :" + uniqueAttribJson);
-                      String id = uniqueAttribJson.getString("id");
-                      String value = uniqueAttribJson.getString("unique-attribute");
-                      String eventType = uniqueAttribJson.getString("eventType");
-                      BroadcastEventType event = BroadcastEventType.from(eventType);
-                      LOGGER.debug("Broadcast event " + event);
-                      if (event == null) {
-                        LOGGER.error("Invalid BroadcastEventType [ null ] ");
-                        return;
-                      }
-                      if (event.equals(BroadcastEventType.CREATE)) {
-                        uniqueAttributeService
-                            .putUniqueAttributeInCache(id, value)
-                            .onSuccess(
-                                successHandler -> {
-                                  LOGGER.debug("unique attrib message published");
-                                  promise.complete();
-                                })
-                            .onFailure(
-                                failureHandler -> {
-                                  LOGGER.debug("unique attrib message published fail");
-                                  promise.fail(
-                                      new ServiceException(
-                                          ERROR_INTERNAL_SERVER, INTERNAL_SERVER_ERROR));
-                                });
-                      }
-                    } else {
-                      LOGGER.info("Empty json received from revoke_token queue");
-                    }
-                  });
-            })
-        .onFailure(
-            failure -> {
-              LOGGER.error("failed ::" + failure.getMessage());
-              promise.fail(new ServiceException(ERROR_INTERNAL_SERVER, INTERNAL_SERVER_ERROR));
-            });
-    return promise.future();
-  }
-}
